requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-doc==0.0.4
    # via typer
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
    #   openai
blinker==1.9.0
    # via flask
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   flask
    #   typer
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tqdm
coverage[toml]==7.13.4
    # via pytest-cov
cryptography==46.0.5
    # via google-auth
distro==1.9.0
    # via
    #   google-genai
    #   openai
flask==3.1.3
    # via -r requirements.in
google-auth[requests]==2.49.0.dev0
    # via google-genai
google-genai==1.65.0
    # via -r requirements.in
gunicorn==25.1.0
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   google-genai
    #   openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.3.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
jiter==0.13.0
    # via openai
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
openai==2.24.0
    # via -r requirements.in
packaging==26.0
    # via
    #   gunicorn
    #   pytest
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pyasn1==0.6.2
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   google-genai
    #   openai
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via
    #   pytest
    #   rich
pytest==9.0.2
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-mock
pytest-cov==7.0.0
    # via -r requirements.in
pytest-mock==3.15.1
    # via -r requirements.in
python-dotenv==1.2.2
    # via -r requirements.in
requests==2.32.5
    # via
    #   google-auth
    #   google-genai
rich==14.3.3
    # via
    #   -r requirements.in
    #   typer
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   google-genai
    #   openai
tenacity==9.1.4
    # via google-genai
tqdm==4.67.3
    # via openai
typer==0.24.1
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   google-genai
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via requests
websockets==16.0
    # via google-genai
werkzeug==3.1.6
    # via flask
